+++ packages/smarthr-ui/src/components/AppHeader/components/mobile/UserInfo.tsx
@@ -82,13 +82,9 @@ const ActualUserInfo: FC<Pick<Props, 'accountUrl'> & { displayName: string }> = 
     <>
       <Dropdown>
         <DropdownTrigger>
-          <Button variant="skeleton" size="s" square className={classNames.iconButton}>
+          <Button variant="skeleton" size="s" className={classNames.iconButton}>
             <span className={classNames.iconButtonInner}>
-              <FaUserLargeIcon
-                role="img"
-                aria-label={translated.account}
-                className="shr-fill-grey"
-              />
+              <FaUserLargeIcon alt={translated.account} className="shr-fill-grey" />
             </span>
           </Button>
         </DropdownTrigger>