+++ packages/smarthr-ui/src/components/SideNav/SideNavItemButton.tsx
@@ -1,4 +1,4 @@
-import { type FC, type ReactNode, memo, useMemo } from 'react'
+import { type ComponentPropsWithoutRef, type FC, type ReactNode, memo, useMemo } from 'react'
 import { tv } from 'tailwind-variants'
 
 import { UnstyledButton } from '../Button'
@@ -9,27 +9,35 @@ export type SideNavSizeType = 'default' | 's'
 type Props = {
   /** アイテムの識別子 */
   id: string
-  /** アイテムのタイトル */
-  title: ReactNode
+  /** アイテムのタイトル
+   * @deprecated SideNav で items を使う時の props です。children を使ってください。
+   */
+  title?: ReactNode
   /** タイトルのプレフィックスの内容。通常、StatusLabel の配置に用います。 */
   prefix?: ReactNode
   /** 選択されているアイテムかどうか */
-  isSelected?: boolean
+  current?: boolean
   /** アイテムの大きさ */
   size?: SideNavSizeType
   /** アイテムを押下したときに発火するコールバック関数 */
   onClick?: (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void
 }
 
+type ElementProps = Omit<ComponentPropsWithoutRef<'li'>, keyof Props>
+
 const classNameGenerator = tv({
   slots: {
     wrapper: [
       'smarthr-ui-SideNav-item',
-      'data-[selected=true]:shr-relative data-[selected=true]:shr-bg-main data-[selected=true]:shr-text-white',
-      'data-[selected=true]:after:shr-absolute data-[selected=true]:after:-shr-right-0.25 data-[selected=true]:after:shr-top-1/2 data-[selected=true]:after:-shr-translate-y-1/2 data-[selected=true]:after:shr-translate-x-0 data-[selected=true]:after:shr-border-b-4 data-[selected=true]:after:shr-border-l-4 data-[selected=true]:after:shr-border-r-0 data-[selected=true]:after:shr-border-t-4 data-[selected=true]:after:shr-border-solid data-[selected=true]:after:shr-border-b-transparent data-[selected=true]:after:shr-border-l-main data-[selected=true]:after:shr-border-r-transparent data-[selected=true]:after:shr-border-t-transparent data-[selected=true]:after:shr-content-[""]',
-      'data-[selected=false]:hover:shr-bg-column-darken',
+      'data-[current=true]:shr-relative data-[current=true]:shr-bg-main data-[current=true]:shr-text-white',
+      'data-[current=true]:after:shr-absolute data-[current=true]:after:-shr-right-0.25 data-[current=true]:after:shr-top-1/2 data-[current=true]:after:-shr-translate-y-1/2 data-[current=true]:after:shr-translate-x-0 data-[current=true]:after:shr-border-b-4 data-[current=true]:after:shr-border-l-4 data-[current=true]:after:shr-border-r-0 data-[current=true]:after:shr-border-t-4 data-[current=true]:after:shr-border-solid data-[current=true]:after:shr-border-b-transparent data-[current=true]:after:shr-border-l-main data-[current=true]:after:shr-border-r-transparent data-[current=true]:after:shr-border-t-transparent data-[current=true]:after:shr-content-[""]',
+      'data-[current=false]:hover:shr-bg-column-darken',
     ],
-    button: ['shr-w-full shr-leading-none [&]:shr-box-border', 'focus-visible:shr-focus-indicator'],
+    button: [
+      'shr-w-full shr-leading-none [&]:shr-box-border',
+      'focus-visible:shr-focus-indicator',
+      'shr-inline-flex shr-items-center',
+    ],
     buttonInner: 'smarthr-ui-SideNav-itemTitle',
   },
   variants: {
@@ -44,21 +52,33 @@ const classNameGenerator = tv({
   },
 })
 
-export const SideNavItemButton: FC<Props> = ({ id, title, prefix, isSelected, size, onClick }) => {
+export const SideNavItemButton: FC<Props & ElementProps> = ({
+  id,
+  title,
+  prefix,
+  current,
+  size,
+  onClick,
+  children,
+}) => {
   const classNames = useMemo(() => {
     const { wrapper, button, buttonInner } = classNameGenerator()
 
     return {
       wrapper: wrapper(),
-      button: button({ size }),
+      button: button({ size: size ?? 'default' }),
       buttonInner: buttonInner(),
     }
   }, [size])
 
   return (
-    <li data-selected={!!isSelected} className={classNames.wrapper}>
+    <li data-current={!!current} className={classNames.wrapper}>
       <UnstyledButton className={classNames.button} onClick={onClick} value={id}>
-        <ButtonBodyCluster prefix={prefix} title={title} titleClassName={classNames.buttonInner} />
+        <ButtonBodyCluster
+          prefix={prefix}
+          title={children ?? title}
+          titleClassName={classNames.buttonInner}
+        />
       </UnstyledButton>
     </li>
   )