+++ packages/smarthr-ui/src/components/Table/Th.tsx
@@ -12,7 +12,7 @@ import { UnstyledButton } from '../Button'
 import { FaSortDownIcon, FaSortUpIcon } from '../Icon'
 import { VisuallyHiddenText } from '../VisuallyHiddenText'
 
-import { reelShadowClassNameGenerator } from './useReelShadow'
+import { reelShadowClassNameGenerator } from './reelShadowStyle'
 
 import type { CellContentWidth } from './type'
 
@@ -23,6 +23,8 @@ export type Props = PropsWithChildren<
     sort?: sortTypes
     /** 並び替えをクリックした時に発火するコールバック関数 */
     onSort?: () => void
+    /** 横スクロール時、カラムを左右いずれかに固定 */
+    fixed?: 'left' | 'right'
     /** 文言を変更するための関数 */
     decorators?: {
       sortDirectionIconAlt: (text: string, { sort }: { sort: sortTypes }) => ReactNode
@@ -63,18 +65,6 @@ const classNameGenerator = tv({
       baseline: 'shr-align-baseline',
       bottom: 'shr-align-bottom',
     },
-    fixed: {
-      left: [
-        /* これ以降の記述はTableReel内で'fixed'を利用した際に追従させるために必要 */
-        '[&.fixed]:shr-sticky [&.fixed]:after:shr-opacity-100',
-        'fixedLeft',
-      ],
-      right: [
-        /* これ以降の記述はTableReel内で'fixed'を利用した際に追従させるために必要 */
-        '[&.fixed]:shr-sticky [&.fixed]:after:shr-opacity-100',
-        'fixedRight',
-      ],
-    },
   },
   defaultVariants: {
     align: 'left',
@@ -108,7 +98,7 @@ export const Th = memo<Props & ElementProps>(
     ...props
   }) => {
     const actualClassName = useMemo(() => {
-      const base = classNameGenerator({ className, align, vAlign, fixed })
+      const base = classNameGenerator({ className, align, vAlign })
 
       if (!fixed) {
         return base
@@ -139,7 +129,13 @@ export const Th = memo<Props & ElementProps>(
     )
 
     return (
-      <th {...props} aria-sort={ariaSort} className={actualClassName} style={actualStyle}>
+      <th
+        {...props}
+        aria-sort={ariaSort}
+        data-fixed={fixed}
+        className={actualClassName}
+        style={actualStyle}
+      >
         {sort ? (
           <MemoizedSortButton align={align} onSort={onSort} sortLabel={sortLabel}>
             {children}
@@ -181,15 +177,17 @@ const MemoizedSortButton = memo<
 
 const sortIconClassNameGenerator = tv({
   slots: {
-    wrapper: 'shr-inline-flex shr-flex-col',
+    wrapper: 'smarthr-ui-Icon-extended shr-relative shr-w-1 shr-h-1',
     upIcon: [
+      'shr-absolute shr-top-0 shr-left-0',
       'shr-text-base',
       '[[aria-sort="none"]_&]:shr-text-disabled',
       '[[aria-sort="ascending"]_&]:shr-text-black',
       '[[aria-sort="descending"]_&]:shr-text-disabled',
     ],
     downIcon: [
-      '-shr-mt-em shr-text-base',
+      'shr-absolute shr-top-0 shr-left-0',
+      'shr-text-base',
       '[[aria-sort="none"]_&]:shr-text-disabled',
       '[[aria-sort="ascending"]_&]:shr-text-disabled',
       '[[aria-sort="descending"]_&]:shr-text-black',